Merge Unified into Master

+++ System.Drawing/Font-CoreText.cs
@@ -28,11 +28,11 @@
 using System;
 
 #if MONOMAC
-using MonoMac.CoreGraphics;
-using MonoMac.CoreText;
+using CoreGraphics;
+using CoreText;
 #else
-using MonoTouch.CoreGraphics;
-using MonoTouch.CoreText;
+using CoreGraphics;
+using CoreText;
 #endif
 
 namespace System.Drawing
@@ -107,9 +107,9 @@ namespace System.Drawing
 			// Documentation for Accessing Font Metrics
 			// http://developer.apple.com/library/ios/#documentation/StringsTextFonts/Conceptual/CoreText_Programming/Operations/Operations.html
 			float lineHeight = 0;
-			lineHeight += nativeFont.AscentMetric;
-			lineHeight += nativeFont.DescentMetric;
-			lineHeight += nativeFont.LeadingMetric;
+			lineHeight += (float)nativeFont.AscentMetric;
+			lineHeight += (float)nativeFont.DescentMetric;
+			lineHeight += (float)nativeFont.LeadingMetric;
 
 
 			// Still have not figured this out yet!!!!
