+++ System.Drawing/Image.cs
@@ -42,13 +42,13 @@ using System.IO;
 using System.Reflection;
 
 #if MONOMAC
-using MonoMac.CoreGraphics;
-using MonoMac.Foundation;
-using MonoMac.AppKit;
+using CoreGraphics;
+using Foundation;
+using AppKit;
 #else
-using MonoTouch.CoreGraphics;
-using MonoTouch.UIKit;
-using MonoTouch.Foundation;
+using CoreGraphics;
+using UIKit;
+using Foundation;
 #endif
 
 namespace System.Drawing {
@@ -92,7 +92,7 @@ namespace System.Drawing {
 		public int Height {
 			get {
 				var b = this as Bitmap;
-				return b == null ? 0 : b.NativeCGImage.Height;
+				return b == null ? 0 : (int)b.NativeCGImage.Height;
 			}
 		}
 		
@@ -117,7 +117,7 @@ namespace System.Drawing {
 		public int Width {
 			get {
 				var b = this as Bitmap;
-				return b == null ? 0 : b.NativeCGImage.Width;
+				return b == null ? 0 : (int)b.NativeCGImage.Width;
 			}
 		}
